+++ OUDS/Core/Components/Sources/Buttons/OUDSButton.swift
@@ -61,7 +61,7 @@ import SwiftUI
 ///
 /// ## Design documentation
 ///
-/// See [unified-design-system.orange.com/472794e18/p/48a788-button](https://unified-design-system.orange.com/472794e18/p/48a788-button)
+/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/48a788-button)
 ///
 /// - Since: 0.10.0
 public struct OUDSButton: View {